# packages/contract/openapi.json
{
  "info": {
    "title": "Animichi Catalog Service",
    "version": "0.1.0",
    "description": "Read methods of the TS Catalog service, consumed by the Python Agent service."
  },
  "openapi": "3.1.1",
  "paths": {
...
    "/catalog/public/popular": {
      "get": {
        "operationId": "popular",
        "summary": "Public popularity ranking: works with published points, rating-descending",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50,
              "default": 8
            },
            "allowEmptyValue": true,
            "allowReserved": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "bangumi": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "bangumi_id": {
                            "type": "string"
                          },
                          "title": {
                            "type": "string"
                          },
                          "title_cn": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "cover_url": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "city": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "points_count": {
                            "type": "integer",
                            "minimum": 0,
                            "maximum": 9007199254740991
                          },
                          "rating": {
                            "anyOf": [
                              {
                                "type": "number"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          }
                        },
                        "required": [
                          "bangumi_id",
                          "title",
                          "title_cn",
                          "cover_url",
                          "city",
                          "points_count",
                          "rating"
                        ]
                      }
                    }
                  },
                  "required": [
                    "bangumi"
                  ]
                }
              }
            }
          }
        }
      }
    }
  }
}